src/scripts/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.1.70
pandas-ta>=0.3.14b
transformers>=4.30.0
requests>=2.31.0
python-dotenv>=1.0.0 